% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{compare_msmah_var}
\alias{compare_msmah_var}
\title{Compare fitted vs true parameters}
\usage{
compare_msmah_var(
  mu_true,
  A_true,
  S_true,
  P_true,
  mu_hat,
  A_hat,
  S_hat,
  P_hat,
  digits = 4
)
}
\arguments{
\item{mu_true}{List of true mean vectors}

\item{A_true}{List of true AR matrices}

\item{S_true}{List of true covariance matrices}

\item{P_true}{True transition matrix}

\item{mu_hat}{List of estimated mean vectors}

\item{A_hat}{List of estimated AR matrices}

\item{S_hat}{List of estimated covariance matrices}

\item{P_hat}{Estimated transition matrix}

\item{digits}{Number of decimal places to print}
}
\description{
Compare fitted vs true parameters
}
