% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmvar_functions.R
\name{dmvnorm_log}
\alias{dmvnorm_log}
\title{Log-density of multivariate normal}
\usage{
dmvnorm_log(x, mean, Sigma, eps = 1e-12)
}
\arguments{
\item{x}{Numeric vector}

\item{mean}{Mean vector}

\item{Sigma}{Covariance matrix}

\item{eps}{Small number to stabilize inversion}
}
\value{
Log-density
}
\description{
Log-density of multivariate normal
}
