% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmvar_functions.R
\name{encode_z}
\alias{encode_z}
\title{Encode augmented state z_t}
\usage{
encode_z(z, M, p)
}
\arguments{
\item{z}{Vector (s_t, s_{t-1},...,s_{t-p})}

\item{M}{Number of regimes}

\item{p}{Lag order}
}
\value{
Integer index
}
\description{
Encode augmented state z_t
}
