% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{MSMAHVAR-package}
\alias{MSMAHVAR}
\alias{MSMAHVAR-package}
\title{MSMAHVAR: Markov-Switching Mean-Adjusted VAR Models}
\description{
Implements EM estimation for MSMAH-VAR(p) models, including:
\itemize{
\item State-dependent means and AR coefficients
\item State-dependent covariances
\item Exact EM for augmented states
\item Fitting to simulated or real datasets
}
}
\author{
\strong{Maintainer}: Angela-Frances Ibhade \email{kiaraibhade@gmail.com}

}
