% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{rmvnorm1}
\alias{rmvnorm1}
\title{Generate one multivariate normal sample}
\usage{
rmvnorm1(mean, Sigma)
}
\arguments{
\item{mean}{Mean vector}

\item{Sigma}{Covariance matrix}
}
\value{
Numeric vector sample
}
\description{
Generate one multivariate normal sample
}
