% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{simulate_msmah_var}
\alias{simulate_msmah_var}
\title{Simulate MSMAH-VAR(p) data}
\usage{
simulate_msmah_var(T, M, k, p, mu, A_list, Sigma_list, P, pi = NULL)
}
\arguments{
\item{T}{Number of time points}

\item{M}{Number of regimes}

\item{k}{Number of variables}

\item{p}{VAR lag order}

\item{mu}{List of state-dependent means}

\item{A_list}{List of state-dependent AR matrices}

\item{Sigma_list}{List of state-dependent covariances}

\item{P}{Transition matrix}

\item{pi}{Initial regime probabilities (default uniform)}
}
\value{
List with Y (data) and s (hidden regimes)
}
\description{
Simulate MSMAH-VAR(p) data
}
