% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{compare_msmah_var}
\alias{compare_msmah_var}
\title{Compare fitted vs true parameters}
\usage{
compare_msmah_var(
  mu_true,
  A_true,
  S_true,
  P_true,
  mu_hat,
  A_hat,
  S_hat,
  P_hat,
  digits = 4
)
}
\description{
Compare fitted vs true parameters
}
