% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{decode_z}
\alias{decode_z}
\title{Decode augmented state index}
\usage{
decode_z(idx, M, p)
}
\arguments{
\item{idx}{Integer index}

\item{M}{Number of regimes}

\item{p}{Lag order}
}
\value{
Vector (s_t, s_{t-1},...,s_{t-p})
}
\description{
Decode augmented state index
}
