% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmahvar_functions.R
\name{logsumexp}
\alias{logsumexp}
\title{Log-sum-exp trick for numerical stability}
\usage{
logsumexp(v)
}
\arguments{
\item{v}{Numeric vector}
}
\value{
log(sum(exp(v)))
}
\description{
Log-sum-exp trick for numerical stability
}
